Rename train helper script to retrain
