+++ docs/index.md
@@ -46,7 +46,7 @@ remain available through Ollama running on the host.
 | Command | Purpose |
 |--------|---------|
 | `agentomics-run` | Run the full agent workflow |
-| `agentomics-train` | Re-train a model with new data |
+| `agentomics-retrain` | Re-train a model with new data |
 | `agentomics-inference` | Run predictions on new data |
 | `agentomics-check-dataset` | Validate a dataset's format before a run |
