+++ pyproject.toml
@@ -34,7 +34,7 @@ Repository = "https://github.com/BioGeMT/Agentomics-ML"
 agentomics-download-dataset = "agentomics.cli.download_dataset:main"
 agentomics-run = "agentomics.cli.run:main"
 agentomics-inference = "agentomics.cli.inference:main"
-agentomics-train = "agentomics.cli.train:main"
+agentomics-retrain = "agentomics.cli.retrain:main"
 agentomics-check-dataset = "agentomics.cli.check_dataset:main"
 
 [tool.setuptools.dynamic]
